# backend/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Technical Challenge",
        "title": "Swagger Technical Challenge API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Luis Enrique Ramírez López",
            "url": "http://www.swagger.io/support",
            "email": "luian.ramirez.12@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
        "/bonds": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a bond",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bonds"
                ],
                "summary": "Create a bond",
                "parameters": [
                    {
                        "description": "Bond object",
                        "name": "bond",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repositories.CreateBondRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repositories.CreateBondResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response400WithResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response404WithResult"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.Response409WithResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response500WithResult"
                        }
                    }
                }
            }
        },
        "/bonds/{type}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all bonds",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bonds"
                ],
                "summary": "Get all bonds",
                "parameters": [
                    {
                        "enum": [
                            "SOLD",
                            "AVAILABLE",
                            "BOUGHT"
                        ],
                        "type": "string",
                        "description": "Type of bond",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repositories.GetAllBondsResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response400WithResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response404WithResult"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.Response409WithResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response500WithResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Response400WithResult": {
            "type": "object",
            "properties": {
                "detalles": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "format": "string"
                    },
                    "example": [
                        "Error en la petición"
                    ]
                },
                "folio": {
                    "type": "string",
                    "format": "string",
                    "example": ""
                },
                "mensaje": {
                    "type": "string",
                    "format": "string",
                    "example": "Operación realizada sin éxito"
                }
            }
        },
        "models.Response404WithResult": {
            "type": "object",
            "properties": {
                "detalles": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "format": "string"
                    },
                    "example": [
                        "No se encontró el recurso solicitado"
                    ]
                },
                "folio": {
                    "type": "string",
                    "format": "string",
                    "example": ""
                },
                "mensaje": {
                    "type": "string",
                    "format": "string",
                    "example": "Operación realizada con éxito"
                }
            }
        },
        "models.Response409WithResult": {
            "type": "object",
            "properties": {
                "folio": {
                    "type": "string",
                    "format": "string",
                    "example": ""
                },
                "mensaje": {
                    "type": "string",
                    "format": "string",
                    "example": "Operación realizada sin éxito"
                }
            }
        },
        "models.Response500WithResult": {
            "type": "object",
            "properties": {
                "detalles": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "format": "string"
                    },
                    "example": [
                        "Error interno del servidor"
                    ]
                },
                "folio": {
                    "type": "string",
                    "format": "string",
                    "example": ""
                },
                "mensaje": {
                    "type": "string",
                    "format": "string",
                    "example": "Operación fallida"
                }
            }
        },
        "repositories.BondModel": {
            "type": "object",
            "properties": {
                "available_quantity": {
                    "type": "integer",
                    "example": 1
                },
                "bond_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "name"
                },
                "price": {
                    "type": "number",
                    "example": 1
                },
                "seller": {
                    "type": "string",
                    "format": "uuid",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "total_quantity": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "repositories.CreateBondRequest": {
            "type": "object",
            "required": [
                "name",
                "price",
                "quantity"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 40,
                    "minLength": 3,
                    "example": "name"
                },
                "price": {
                    "type": "number",
                    "maximum": 100000000,
                    "minimum": 0.0001,
                    "example": 1
                },
                "quantity": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 1,
                    "example": 1
                }
            }
        },
        "repositories.CreateBondResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "repositories.CreateBondResponse200": {
            "type": "object",
            "properties": {
                "folio": {
                    "type": "string",
                    "format": "string",
                    "example": ""
                },
                "mensaje": {
                    "type": "string",
                    "format": "string",
                    "example": "Operación realizada con éxito"
                },
                "result": {
                    "$ref": "#/definitions/repositories.CreateBondResponse"
                }
            }
        },
        "repositories.GetAllBondsResponse": {
            "type": "object",
            "properties": {
                "bonds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repositories.BondModel"
                    }
                }
            }
        },
        "repositories.GetAllBondsResponse200": {
            "type": "object",
            "properties": {
                "folio": {
                    "type": "string",
                    "format": "string",
                    "example": ""
                },
                "mensaje": {
                    "type": "string",
                    "format": "string",
                    "example": "Operación realizada con éxito"
                },
                "result": {
                    "$ref": "#/definitions/repositories.GetAllBondsResponse"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}